add: トロフィー機能の修正・Preferences系の修正をしました
