+++ lib/ui/screen/trophy_screen/children/check_text.dart
